Make MinidumpStream::STREAM_TYPE a u32
Fixes #362

+++ minidump-stackwalk/src/main.rs
@@ -668,7 +668,7 @@ where
         LinuxCpuInfo,
         LinuxMaps
     ) {
-        if let Ok(contents) = dump.get_raw_stream(stream) {
+        if let Ok(contents) = dump.get_raw_stream(stream as u32) {
             print_raw_stream(name, contents, output)?;
         }
     }
